Refactoring code: split ShopVMContext into modules. Fixing scrollbars and modals.
