+++ src/models/map/types.ts
@@ -1,0 +1,16 @@
+export type Coordinates = {
+  latitude: number
+  longitude: number
+}
+
+export type Store = {
+  id: string
+  name: string
+  street: string
+  housenumber: string
+  postcode: string
+  city: string
+  brand: string
+  branch: string
+  coordinates: Coordinates
+}
