+++ src/viewmodels/ShopVMContext.tsx
@@ -1,9 +1,54 @@
 // src/viewmodels/ShopVMContext.tsx
-// YHTENÄINEN malli:
-// - Kaikki listat: /lists/{listId}
-// - Kategoriat: /lists/{listId}/categories
-// - Itemit: /lists/{listId}/items
-// - Järjestys per käyttäjä: list.orderBy[uid]
+
+// Tämä on sovelluksen ViewModel/Context: yksi paikka, josta UI saa
+// ostoslistoihin liittyvän tilan ja toiminnot.
+//
+// UI saa täältä:
+// - käyttäjän (Firebase Auth) ja uid:n
+// - stores: käyttäjän omat “kaupat” (users/{uid}/stores)
+// - lists: jaettavat listat (/lists, jäsenyys memberIds-taulukosta)
+// - categories + items: listan alakokoelmista (/lists/{listId}/categories ja /items)
+//
+// Tärkeä periaate tässä tiedostossa:
+// - ShopVMContext sisältää UI:n kannalta tärkeät “wrapperit” ja state-päivitykset
+// - Varsinaiset Firestore-operaatiot on siirretty pienempiin shop/* tiedostoihin
+
+//FIRESTORE-MALLI (rakenne):
+// 
+// 1) Listat (yhteiset, jaettavat):
+//    /lists/{listId}
+//       - name: listan nimi
+//       - storeId: (valinnainen) viite käyttäjän omaan storeen
+//       - ownerId: listan omistajan uid
+//       - memberIds: uid-taulukko -> ketkä näkee listan
+//       - orderBy: { [uid]: number } -> jokaisella käyttäjällä oma listajärjestys
+//
+// 2) Kategoriat (listakohtaiset):
+//    /lists/{listId}/categories/{categoryId}
+//       - name: kategorian nimi
+//       - order: järjestysnumero listan sisällä
+//
+// 3) Itemit (listakohtaiset):
+//    /lists/{listId}/items/{itemId}
+//       - name: tuotteen nimi
+//       - done: onko “ruksattu”
+//       - categoryId: mihin kategoriaan kuuluu (tai null = ei kategoriaa)
+//       - order: järjestysnumero kategorian/ listan sisällä
+//       - quantity: määrä (min 1)
+//
+// 4) Kutsut (invites):
+//    /invites/{CODE}
+//       - listId: mihin listaan kutsu liittyy
+//       - createdBy: kuka loi kutsun
+//
+// 
+// (HUOMIO: stores on edelleen käyttäjäkohtainen)
+// 
+// /users/{uid}/stores/{storeId}
+//  - name, branch
+//
+// Tällä mallilla listat ovat jaettavia (/lists), mutta “kauppavalinnat” voivat
+// olla käyttäjäkohtaisia (stores). StoreId on vain string-viite listassa.
 
 import React, { createContext, useContext, useEffect, useMemo, useRef, useState } from "react"
 import type { User } from "firebase/auth"
@@ -11,177 +56,85 @@ import type { User } from "firebase/auth"
 import {
   auth,
   db,
+
+  // Auth-toiminnot
   onAuthStateChanged,
   signInAnonymously,
-  collection,
-  addDoc,
-  onSnapshot,
-  query,
-  orderBy,
-  serverTimestamp,
-  getDoc,
-  setDoc,
-  getDocs,
-  deleteDoc,
-  updateDoc,
-  doc,
-  writeBatch,
-  where,
-  runTransaction,
-  arrayUnion,
 } from "../../firebase/Config"
 
-export type Store = {
-  id: string
-  name: string
-  branch?: string
-}
+// Tyypit on eriytetty omaan tiedostoon (shop/types.ts),
+// jotta ViewModel ei paisu pelkistä type-määrittelyistä.
+import type { VM, Store, ShopList, Category, ListItem } from "./shop/types"
+export type { Store, ShopList, Category, ListItem } from "./shop/types"
 
-export type ShopList = {
-  id: string
-  name: string
-  storeId: string | null
-  order: number
-  ownerId: string
-  memberIds: string[]
-}
+// Seed (oletuskategoriat) ja kutsulogiikka omissa tiedostoissaan.
+import { ensureDefaultListCategories } from "./shop/seed"
+import { createInviteCodeForList as createInviteCodeForListSvc, joinListByCode as joinListByCodeSvc } from "./shop/invites"
 
-export type Category = {
-  id: string
-  name: string
-  order: number
-}
+// Listojen Firestore-operaatiot omassa tiedostossa.
+import { createListDoc, deleteListDeep, reorderListsForUser } from "./shop/lists"
 
-export type ListItem = {
-  id: string
-  name: string
-  done: boolean
-  categoryId: string | null
-  order: number
-  quantity: number
-}
+// Itemien Firestore-operaatiot omassa tiedostossa.
+import {
+  subscribeItems as subscribeItemsSvc,
+  addItemDoc,
+  updateItemDoc,
+  deleteItemDoc,
+  reorderItemsInCategoryDb,
+  setItemQuantity,
+} from "./shop/items"
 
-type StoreDoc = {
-  name: string
-  branch?: string | null
-  createdAt?: unknown
-}
+// Kategorioiden Firestore-operaatiot omassa tiedostossa.
+import { subscribeCategories as subscribeCategoriesSvc, addCategoryDoc, reorderCategoriesDb } from "./shop/categories"
 
-type ListDoc = {
-  name: string
-  storeId?: string | null
-  ownerId: string
-  memberIds: string[]
-  orderBy: Record<string, number>
-  createdAt?: unknown
-}
+// Stores-toiminnot omassa tiedostossa.
+import { createStoreDoc, deleteStoreDoc } from "./shop/stores"
 
-type CategoryDoc = {
-  name: string
-  order: number
-  createdAt?: unknown
-}
+// Reaaliaikaiset kuuntelut (stores + lists) omassa tiedostossa.
+import { subscribeStores, subscribeLists } from "./shop/subscriptions"
 
-type ItemDoc = {
-  name: string
-  done: boolean
-  categoryId: string | null
-  order: number
-  quantity: number
-  createdAt?: unknown
-}
-
-type InviteDoc = {
-  listId: string
-  createdBy: string
-  createdAt?: unknown
-}
-
-const DEFAULT_LIST_CATEGORIES = [
-  "Hedelmät & vihannekset",
-  "Maito & kananmunat",
-  "Leipä",
-  "Liha & kala",
-  "Valmisruoka",
-  "Kuivat tuotteet",
-  "Pakasteet",
-  "Juomat",
-  "Makeiset & naposteltavat",
-  "Hygienia",
-  "Koti & siivous",
-]
-
-type VM = {
-  // Auth
-  user: User | null
-  uid: string | null
-
-  // Data
-  stores: Store[]
-  lists: ShopList[]
-  categoriesByListId: Record<string, Category[]>
-  itemsByListId: Record<string, ListItem[]>
-
-  // Stores (pidetään ennallaan users/{uid}/stores)
-  createStore: (name: string, branch?: string) => Promise<void>
-  deleteStore: (storeId: string) => Promise<void>
-
-  // Lists (uusi malli /lists)
-  createList: (name: string, storeId: string | null) => Promise<string | null>
-  deleteList: (listId: string) => Promise<void>
-  reorderLists: (nextListIds: string[]) => Promise<void>
-
-  // Invites
-  createInviteCodeForList: (listId: string) => Promise<string | null>
-  joinListByCode: (code: string) => Promise<string | null>
-
-  // Categories
-  subscribeCategoriesForList: (listId: string) => () => void
-  createCategoryForList: (listId: string, name: string) => Promise<void>
-  reorderCategoriesForList: (listId: string, nextIds: string[]) => Promise<void>
-
-  // Items
-  subscribeItems: (listId: string) => () => void
-  addItem: (listId: string, name: string, categoryId: string | null) => Promise<void>
-  updateItem: (
-    listId: string,
-    itemId: string,
-    patch: Partial<Pick<ListItem, "name" | "done" | "categoryId" | "order" | "quantity">>
-  ) => Promise<void>
-  deleteItem: (listId: string, itemId: string) => Promise<void>
-  reorderItemsInCategory: (listId: string, categoryId: string | null, nextItemIds: string[]) => Promise<void>
-  changeQuantity: (listId: string, itemId: string, delta: number) => Promise<void>
-
-  // Helpers
-  getStoreLabel: (storeId: string | null) => string | undefined
-  isOwnerOfList: (list: ShopList) => boolean
-}
-
+// Context sisältää VM:n. Ilman Provideria arvo on null -> hook heittää selkeän virheen.
 const Ctx = createContext<VM | null>(null)
 
 export function ShopVMProvider({ children }: { children: React.ReactNode }) {
-
+  // Auth state:
+  // - user: Firebase Authin User-objekti (voi olla null)
+  // - uid: user.uid erikseen stringinä, jotta sitä on helppo käyttää kaikkialla
   const [user, setUser] = useState<User | null>(null)
   const [uid, setUid] = useState<string | null>(null)
 
+  // Päädata:
+  // - stores: käyttäjän omat kaupat (users/{uid}/stores)
+  // - lists: listat joihin käyttäjä kuuluu (/lists where memberIds contains uid)
   const [stores, setStores] = useState<Store[]>([])
   const [lists, setLists] = useState<ShopList[]>([])
 
+  // Listakohtainen data mappeina:
+  // - categoriesByListId[listId] = Category[]
+  // - itemsByListId[listId] = ListItem[]
+  //
+  // Map-rakenne on nopea: UI saa suoraan “tämän listan” datan ilman suodatusta.
   const [categoriesByListId, setCategoriesByListId] = useState<Record<string, Category[]>>({})
   const [itemsByListId, setItemsByListId] = useState<Record<string, ListItem[]>>({})
 
-  // jotta ei seedailla samaa listaa useasti saman session aikana
+  // Seedaus-esto (vain session ajaksi):
+  // Ref on tärkeä, koska:
+  // - se ei aiheuta rerenderiä
+  // - se säilyy elossa Providerin elinkaaren ajan
   const seededListIdsRef = useRef<Set<string>>(new Set())
 
+  // Apuri: poistaa listatuplikaatit id:n perusteella.
+  // Tätä tarvitaan, koska createList tekee optimistisen lisäyksen, ja sen jälkeen
+  // snapshot voi tuoda saman listan uudestaan -> hetkellinen duplikaatti.
   const dedupeLists = (arr: ShopList[]) => {
     const m = new Map<string, ShopList>()
     for (const l of arr) m.set(l.id, l)
     return Array.from(m.values())
-}
+  }
 
-  console.log("UID:", uid)
-
-  // AUTH
+  // AUTH:
+  // - kuunnellaan kirjautumistilaa
+  // - jos käyttäjää ei ole, kirjaudutaan anonyymisti
   useEffect(() => {
     const unsub = onAuthStateChanged(auth, async (u) => {
       if (u) {
@@ -202,82 +155,26 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
     return unsub
   }, [])
 
-  // SNAPSHOTS: stores + lists
+  // SNAPSHOTS:
+  // Kun uid on tiedossa, käynnistetään reaaliaikaiset kuuntelut:
+  // - stores: käyttäjän omat kaupat
+  // - lists: listat, joissa käyttäjä on jäsen
   useEffect(() => {
     if (!uid) return
 
-    // STORES: users/{uid}/stores (tätä voi käyttää edelleen “kauppavalintana”)
-    const storesRef = collection(db, "users", uid, "stores")
-    const storesQ = query(storesRef, orderBy("createdAt", "desc"))
-
-    const unsubStores = onSnapshot(storesQ, (snap) => {
-      const next: Store[] = snap.docs
-        .map((d) => {
-          const data = d.data() as Partial<StoreDoc>
-          if (typeof data.name !== "string") return null
-          return {
-            id: d.id,
-            name: data.name,
-            ...(typeof data.branch === "string" ? { branch: data.branch } : {}),
-          }
-        })
-        .filter((x): x is Store => x !== null)
-
-      setStores(next)
+    const unsubStores = subscribeStores({
+      db,
+      uid,
+      onStores: (next) => setStores(next),
+      onError: (err) => console.error("Stores onSnapshot error:", err),
     })
 
-    // LISTS: /lists where memberIds contains uid
-    const listsRef = collection(db, "lists")
-    const listsQ = query(listsRef, where("memberIds", "array-contains", uid))
-
-    // DEBUG: testaa query kertahaulla
-    getDocs(listsQ)
-      .then((snapOnce) => {
-        console.log(
-          "DEBUG getDocs lists count:",
-          snapOnce.size,
-          snapOnce.docs.map((d) => d.id)
-        )
-      })
-      .catch((e) => console.error("DEBUG getDocs error:", e))
-
-
-    const unsubLists = onSnapshot(
-      listsQ,
-      (snap) => {
-        const next = snap.docs
-          .map((d) => {
-            const data = d.data() as Partial<ListDoc>
-            if (typeof data.name !== "string") return null
-            if (typeof data.ownerId !== "string") return null
-            if (!Array.isArray(data.memberIds)) return null
-
-            const storeId =
-              typeof data.storeId === "string"
-                ? data.storeId
-                : data.storeId === null
-                  ? null
-                  : null
-
-            const orderByMap = (data.orderBy ?? {}) as Record<string, number>
-            const order = typeof orderByMap[uid] === "number" ? orderByMap[uid] : 999999
-
-            return {
-              id: d.id,
-              name: data.name,
-              storeId,
-              order,
-              ownerId: data.ownerId,
-              memberIds: data.memberIds.filter((x): x is string => typeof x === "string"),
-            } satisfies ShopList
-          })
-          .filter((x): x is ShopList => x !== null)
-          .sort((a, b) => a.order - b.order)
-
-        setLists(dedupeLists(next))
-      },
-      (err) => console.error("Lists onSnapshot error:", err)
-    )
+    const unsubLists = subscribeLists({
+      db,
+      uid,
+      onLists: (next) => setLists(dedupeLists(next)),
+      onError: (err) => console.error("Lists onSnapshot error:", err),
+    })
 
     return () => {
       unsubStores()
@@ -285,189 +182,80 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
     }
   }, [uid])
 
-  /* STORES actions */
+  /* STORES */
 
   const createStore = async (name: string, branch?: string) => {
     if (!uid) return
     const trimmed = name.trim()
     if (!trimmed) return
-
-    await addDoc(collection(db, "users", uid, "stores"), {
-      name: trimmed,
-      branch: branch?.trim() ?? null,
-      createdAt: serverTimestamp(),
-    } satisfies StoreDoc)
+    await createStoreDoc({ db, uid, name: trimmed, branch })
   }
 
   const deleteStore = async (storeId: string) => {
     if (!uid) return
-    await deleteDoc(doc(db, "users", uid, "stores", storeId))
+    await deleteStoreDoc({ db, uid, storeId })
   }
 
-  /* LIST helpers */
-
-  const ensureDefaultListCategories = async (listId: string) => {
-    if (!uid) return
-    if (seededListIdsRef.current.has(listId)) return
-
-    const ref = collection(db, "lists", listId, "categories")
-    const snap = await getDocs(ref)
-    if (!snap.empty) {
-      seededListIdsRef.current.add(listId)
-      return
-    }
-
-    const batch = writeBatch(db)
-    DEFAULT_LIST_CATEGORIES.forEach((name, index) => {
-      const newDoc = doc(ref)
-      batch.set(newDoc, { name, order: index, createdAt: serverTimestamp() })
-    })
-    await batch.commit()
-
-    seededListIdsRef.current.add(listId)
-  }
-
-  /* LISTS actions */
+  /* LISTS */
 
   const createList = async (name: string, storeId: string | null) => {
     if (!uid) return null
+
     const trimmed = name.trim()
     if (!trimmed) return null
 
+    // Käyttäjäkohtainen järjestysnumero tälle uudelle listalle.
+    // Jos listoja ei ole, currentMax = -1 -> nextOrder = 0.
     const currentMax = lists.reduce((m, l) => Math.max(m, l.order), -1)
     const nextOrder = currentMax + 1
 
-    const ref = await addDoc(collection(db, "lists"), {
+    // 1) Luodaan listadokumentti Firestoreen
+    const listId = await createListDoc({
+      db,
+      uid,
       name: trimmed,
       storeId: storeId ?? null,
-      ownerId: uid,
-      memberIds: [uid],
-      orderBy: { [uid]: nextOrder },
-      createdAt: serverTimestamp(),
-    } satisfies ListDoc)
+      orderForUser: nextOrder,
+    })
 
-    // lisätään lista heti lokaalisti, ettei /shoplist ehdi näyttää "ei löydy"
+    // 2) Optimistinen UI: lisätään lista heti stateen
     setLists((prev) => {
       const merged = dedupeLists([
         ...prev,
-        {
-          id: ref.id,
-          name: trimmed,
-          storeId: storeId ?? null,
-          order: nextOrder,
-          ownerId: uid,
-          memberIds: [uid],
-        },
+        { id: listId, name: trimmed, storeId: storeId ?? null, order: nextOrder, ownerId: uid, memberIds: [uid] },
       ])
       return merged.sort((a, b) => a.order - b.order)
     })
 
-    await ensureDefaultListCategories(ref.id)
-    return ref.id
+    // 3) Seedaus: oletuskategoriat, jos listalla ei ole vielä yhtään kategoriaa
+    await ensureDefaultListCategories(db, listId, seededListIdsRef.current)
+
+    return listId
   }
 
   const deleteList = async (listId: string) => {
     if (!uid) return
-
-    // Poista items
-    const itemsRef = collection(db, "lists", listId, "items")
-    const itemsSnap = await getDocs(itemsRef)
-    const itemDocs = itemsSnap.docs
-
-    // Poista categories
-    const catsRef = collection(db, "lists", listId, "categories")
-    const catsSnap = await getDocs(catsRef)
-    const catDocs = catsSnap.docs
-
-    const chunkSize = 450
-
-    for (let i = 0; i < itemDocs.length; i += chunkSize) {
-      const batch = writeBatch(db)
-      itemDocs.slice(i, i + chunkSize).forEach((d) => batch.delete(d.ref))
-      await batch.commit()
-    }
-
-    for (let i = 0; i < catDocs.length; i += chunkSize) {
-      const batch = writeBatch(db)
-      catDocs.slice(i, i + chunkSize).forEach((d) => batch.delete(d.ref))
-      await batch.commit()
-    }
-
-    await deleteDoc(doc(db, "lists", listId))
+    await deleteListDeep({ db, listId })
   }
 
   const reorderLists = async (nextListIds: string[]) => {
     if (!uid) return
-
-    const batch = writeBatch(db)
-    nextListIds.forEach((id, idx) => {
-      batch.update(doc(db, "lists", id), {
-        [`orderBy.${uid}`]: idx,
-      })
-    })
-    await batch.commit()
+    await reorderListsForUser({ db, uid, nextListIds })
   }
 
   /* INVITES */
 
-  const makeCode = () => Math.random().toString(36).slice(2, 8).toUpperCase()
-
+  // Wrapperit:
+  // - varmistetaan uid
+  // - delegoidaan Firestore-työ invites.ts tiedostolle
   const createInviteCodeForList = async (listId: string) => {
     if (!uid) return null
-
-    // varmistus: olet jäsen
-    const listRef = doc(db, "lists", listId)
-    const listSnap = await getDoc(listRef)
-    if (!listSnap.exists()) return null
-    const list = listSnap.data() as Partial<ListDoc>
-    if (!Array.isArray(list.memberIds) || !list.memberIds.includes(uid)) return null
-
-    for (let i = 0; i < 5; i++) {
-      const code = makeCode()
-      const inviteRef = doc(db, "invites", code)
-      const existing = await getDoc(inviteRef)
-      if (existing.exists()) continue
-
-      await setDoc(inviteRef, {
-        listId,
-        createdBy: uid,
-        createdAt: serverTimestamp(),
-      } satisfies InviteDoc)
-
-      return code
-    }
-
-    return null
+    return createInviteCodeForListSvc(db, uid, listId)
   }
 
-  const joinListByCode = async (codeRaw: string) => {
+  const joinListByCode = async (code: string) => {
     if (!uid) return null
-
-    const code = codeRaw.trim().toUpperCase()
-    if (!code) return null
-
-    const listId = await runTransaction(db, async (tx) => {
-      const inviteRef = doc(db, "invites", code)
-      const inviteSnap = await tx.get(inviteRef)
-      if (!inviteSnap.exists()) throw new Error("Kutsua ei löydy")
-
-      const inv = inviteSnap.data() as Partial<InviteDoc>
-      if (typeof inv.listId !== "string") throw new Error("Kutsu on virheellinen")
-
-      const listRef = doc(db, "lists", inv.listId)
-
-      tx.update(listRef, {
-        memberIds: arrayUnion(uid),
-        [`orderBy.${uid}`]: 999999,
-      })
-
-      // siisti: poista kutsu käytön jälkeen
-      tx.delete(inviteRef)
-
-      return inv.listId
-    })
-
-    return listId
+    return joinListByCodeSvc(db, uid, code)
   }
 
   /* CATEGORIES */
@@ -475,23 +263,14 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
   const subscribeCategoriesForList = (listId: string) => {
     if (!uid) return () => { }
 
-    // varmistetaan defaultit (vain kerran / session)
-    ensureDefaultListCategories(listId)
+    // Varmistetaan oletuskategoriat (vain kerran per lista / sessio)
+    ensureDefaultListCategories(db, listId, seededListIdsRef.current)
 
-    const ref = collection(db, "lists", listId, "categories")
-    const q = query(ref, orderBy("order", "asc"))
-
-    const unsub = onSnapshot(q, (snap) => {
-      const next: Category[] = snap.docs
-        .map((d) => {
-          const data = d.data() as Partial<CategoryDoc>
-          if (typeof data.name !== "string") return null
-          const order = typeof data.order === "number" ? data.order : 0
-          return { id: d.id, name: data.name, order }
-        })
-        .filter((x): x is Category => x !== null)
-
-      setCategoriesByListId((prev) => ({ ...prev, [listId]: next }))
+    const unsub = subscribeCategoriesSvc({
+      db,
+      listId,
+      onCategories: (next) => setCategoriesByListId((prev) => ({ ...prev, [listId]: next })),
+      onError: (err) => console.error("Categories onSnapshot error:", err),
     })
 
     return () => {
@@ -513,17 +292,13 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
     const maxOrder = current.reduce((m, c) => Math.max(m, c.order), -1)
     const nextOrder = maxOrder + 1
 
-    await addDoc(collection(db, "lists", listId, "categories"), {
-      name: trimmed,
-      order: nextOrder,
-      createdAt: serverTimestamp(),
-    } satisfies CategoryDoc)
+    await addCategoryDoc({ db, listId, name: trimmed, order: nextOrder })
   }
 
   const reorderCategoriesForList = async (listId: string, nextIds: string[]) => {
     if (!uid) return
 
-    // optimistinen UI
+    // Optimistinen UI: päivitetään orderit lokaaliin stateen heti
     setCategoriesByListId((prev) => {
       const current = prev[listId] ?? []
       const byId = new Map(current.map((c) => [c.id, c]))
@@ -536,15 +311,7 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
       return { ...prev, [listId]: next }
     })
 
-    const chunkSize = 450
-    for (let i = 0; i < nextIds.length; i += chunkSize) {
-      const batch = writeBatch(db)
-      nextIds.slice(i, i + chunkSize).forEach((id, localIndex) => {
-        const order = i + localIndex
-        batch.update(doc(db, "lists", listId, "categories", id), { order })
-      })
-      await batch.commit()
-    }
+    await reorderCategoriesDb({ db, listId, nextIds })
   }
 
   /* ITEMS */
@@ -552,32 +319,11 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
   const subscribeItems = (listId: string) => {
     if (!uid) return () => { }
 
-    const ref = collection(db, "lists", listId, "items")
-    const q = query(ref, orderBy("order", "asc"))
-
-    const unsub = onSnapshot(q, (snap) => {
-      const next: ListItem[] = snap.docs
-        .map((d) => {
-          const data = d.data() as Partial<ItemDoc>
-          if (typeof data.name !== "string") return null
-
-          return {
-            id: d.id,
-            name: data.name,
-            done: !!data.done,
-            categoryId:
-              typeof data.categoryId === "string"
-                ? data.categoryId
-                : data.categoryId === null
-                  ? null
-                  : null,
-            order: typeof data.order === "number" ? data.order : 0,
-            quantity: typeof data.quantity === "number" ? data.quantity : 1,
-          }
-        })
-        .filter((x): x is ListItem => x !== null)
-
-      setItemsByListId((prev) => ({ ...prev, [listId]: next }))
+    const unsub = subscribeItemsSvc({
+      db,
+      listId,
+      onItems: (next) => setItemsByListId((prev) => ({ ...prev, [listId]: next })),
+      onError: (err) => console.error("Items onSnapshot error:", err),
     })
 
     return () => {
@@ -592,7 +338,6 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
 
   const addItem = async (listId: string, name: string, categoryId: string | null) => {
     if (!uid) return
-
     const trimmed = name.trim()
     if (!trimmed) return
 
@@ -600,14 +345,7 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
     const maxOrder = current.reduce((m, it) => Math.max(m, it.order), -1)
     const nextOrder = maxOrder + 1
 
-    await addDoc(collection(db, "lists", listId, "items"), {
-      name: trimmed,
-      done: false,
-      categoryId: categoryId ?? null,
-      order: nextOrder,
-      quantity: 1,
-      createdAt: serverTimestamp(),
-    } satisfies ItemDoc)
+    await addItemDoc({ db, listId, name: trimmed, categoryId, order: nextOrder })
   }
 
   const updateItem = async (
@@ -617,42 +355,46 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
   ) => {
     if (!uid) return
 
-    const out: Partial<ItemDoc> = {}
+    // Firestoreen ei kannata lähettää undefined-arvoja,
+    // joten rakennetaan patch puhtaaksi.
+    const out: any = {}
     if (patch.name !== undefined) out.name = patch.name
     if (patch.done !== undefined) out.done = patch.done
     if (patch.categoryId !== undefined) out.categoryId = patch.categoryId
     if (patch.order !== undefined) out.order = patch.order
     if (patch.quantity !== undefined) out.quantity = patch.quantity
 
-    await updateDoc(doc(db, "lists", listId, "items", itemId), out)
+    await updateItemDoc({ db, listId, itemId, patch: out })
   }
 
   const deleteItem = async (listId: string, itemId: string) => {
     if (!uid) return
-    await deleteDoc(doc(db, "lists", listId, "items", itemId))
+    await deleteItemDoc({ db, listId, itemId })
   }
 
   const changeQuantity = async (listId: string, itemId: string, delta: number) => {
     if (!uid) return
+
     const current = itemsByListId[listId]?.find((i) => i.id === itemId)
     const next = Math.max(1, (current?.quantity ?? 1) + delta)
 
-    // optimistinen UI
+    // Optimistinen UI: päivitys näkyy heti
     setItemsByListId((prev) => ({
       ...prev,
       [listId]: (prev[listId] ?? []).map((it) => (it.id === itemId ? { ...it, quantity: next } : it)),
     }))
 
-    await updateDoc(doc(db, "lists", listId, "items", itemId), { quantity: next })
+    await setItemQuantity({ db, listId, itemId, quantity: next })
   }
 
   const reorderItemsInCategory = async (listId: string, categoryId: string | null, nextItemIds: string[]) => {
     if (!uid) return
 
-    // optimistinen UI
+    // Optimistinen UI: rakennetaan uusi järjestys vain tälle kategorialle
     setItemsByListId((prev) => {
       const all = prev[listId] ?? []
       const catKey = categoryId ?? null
+
       const inCat = all.filter((it) => it.categoryId === catKey)
       const outCat = all.filter((it) => it.categoryId !== catKey)
 
@@ -667,15 +409,7 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
       return { ...prev, [listId]: [...outCat, ...nextInCat] }
     })
 
-    const chunkSize = 450
-    for (let i = 0; i < nextItemIds.length; i += chunkSize) {
-      const batch = writeBatch(db)
-      nextItemIds.slice(i, i + chunkSize).forEach((id, localIndex) => {
-        const order = i + localIndex
-        batch.update(doc(db, "lists", listId, "items", id), { order, categoryId: categoryId ?? null })
-      })
-      await batch.commit()
-    }
+    await reorderItemsInCategoryDb({ db, listId, categoryId, nextItemIds })
   }
 
   /* HELPERS */
@@ -691,6 +425,7 @@ export function ShopVMProvider({ children }: { children: React.ReactNode }) {
     return !!uid && list.ownerId === uid
   }
 
+  // Contextin arvo tallennetaan memoon, jotta UI ei renderöidy turhaan
   const value = useMemo<VM>(
     () => ({
       user,
